Add a rule that suggests adding an index to what appear to be foreign/external IDs by convention
This rule may help rails developers be more conscious of adding indexes. While it's not a hard rule,
columns ending with _id generally signal a foreign key that is used in where clauses or joins. Unless
you are absolutely not querying against that value, you probably want an index.

This is fairly opinionated but should at least be considered each time a migration is created.

Limitations:

* only applies to adding columns, not creating tables
* doesn't detect composite index use. can't suggest composite indexes for columns not included in the same file
* doesn't apply a fix[1]

[1] I tried using a regex and wasn't sure how to include metavariables in fix-regex replacement.
I tried using a pattern but wasn't sure how to account for the optional trailing parameters so they would get erased.